# packages/contracts/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Va de Vi API",
    "version": "0.1.0",
    "description": "Space-scoped API contracts for the private Va de Vi application."
  },
  "components": {
    "securitySchemes": {
      "FirebaseBearer": {
        "bearerFormat": "Firebase ID token",
        "scheme": "bearer",
        "type": "http"
      }
    },
    "schemas": {
...
      "BootstrapResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "features": {
                "type": "object",
                "properties": {
                  "assistant": {
                    "type": "boolean"
                  },
                  "externalResearch": {
                    "type": "boolean"
                  },
                  "priceLookup": {
                    "type": "boolean"
                  },
                  "voiceInput": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "assistant",
                  "externalResearch",
                  "priceLookup",
                  "voiceInput"
                ],
                "additionalProperties": false
              },
              "spaces": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "pattern": "^[0-9A-HJKMNP-TV-Z]{26}$"
                    },
                    "name": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 120
                    },
                    "role": {
                      "type": "string",
                      "enum": [
                        "owner",
                        "admin",
                        "member"
                      ]
                    },
                    "type": {
                      "type": "string",
                      "enum": [
                        "personal",
                        "couple",
                        "group"
                      ]
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "role",
                    "type"
                  ],
                  "additionalProperties": false
                }
              },
              "user": {
                "type": "object",
                "properties": {
                  "activeSpaceId": {
                    "type": "string",
                    "pattern": "^[0-9A-HJKMNP-TV-Z]{26}$"
                  },
                  "displayName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 120
                  },
                  "id": {
                    "type": "string",
                    "pattern": "^[0-9A-HJKMNP-TV-Z]{26}$"
                  },
                  "onboardingComplete": {
                    "type": "boolean"
                  },
                  "preferredLocale": {
                    "type": "string",
                    "enum": [
                      "ca",
                      "es",
                      "fr",
                      "en",
                      "it",
                      "pt-PT",
                      "nl",
                      "de"
                    ]
                  }
                },
                "required": [
                  "activeSpaceId",
                  "displayName",
                  "id",
                  "onboardingComplete",
                  "preferredLocale"
                ],
                "additionalProperties": false
              },
              "versions": {
                "type": "object",
                "properties": {
                  "api": {
                    "type": "string",
                    "minLength": 1
                  },
                  "i18nCatalog": {
                    "type": "string",
                    "minLength": 1
                  },
                  "tastingOntology": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "api",
                  "i18nCatalog",
                  "tastingOntology"
                ],
                "additionalProperties": false
              }
            },
            "required": [
              "features",
              "spaces",
              "user",
              "versions"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "data"
        ],
        "additionalProperties": false
      },
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "$ref": "#/components/schemas/ErrorCode"
              },
              "message": {
                "type": "string",
                "minLength": 1
              },
              "requestId": {
                "type": "string",
                "minLength": 1
              },
              "details": {
                "type": "object",
                "additionalProperties": {
                  "nullable": true
                }
              }
            },
            "required": [
              "code",
              "message",
              "requestId"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
      "ErrorCode": {
        "type": "string",
        "enum": [
          "AUTH_REQUIRED",
          "AUTH_INVALID",
          "FORBIDDEN",
          "NOT_FOUND",
          "VALIDATION_FAILED",
          "VERSION_CONFLICT",
          "IDEMPOTENCY_CONFLICT",
          "RATE_LIMITED",
          "QUOTA_EXHAUSTED",
          "FEATURE_UNAVAILABLE",
          "EXTERNAL_SOURCE_UNAVAILABLE",
          "MEDIA_REJECTED",
          "SYNC_CURSOR_EXPIRED",
          "INVITE_INVALID",
          "INVITE_EXPIRED",
          "INTERNAL_ERROR"
        ]
      },
      "UpdateProfileRequest": {
        "type": "object",
        "properties": {
          "activeSpaceId": {
            "type": "string",
            "pattern": "^[0-9A-HJKMNP-TV-Z]{26}$"
          },
          "completeOnboarding": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "displayName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "preferredLocale": {
            "type": "string",
            "enum": [
              "ca",
              "es",
              "fr",
              "en",
              "it",
              "pt-PT",
              "nl",
              "de"
            ]
          }
        },
        "additionalProperties": false
      },
      "SpaceDetailResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "members": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "displayName": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 120
                    },
                    "id": {
                      "type": "string",
                      "pattern": "^[0-9A-HJKMNP-TV-Z]{26}$"
                    },
                    "joinedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "role": {
                      "type": "string",
                      "enum": [
                        "owner",
                        "admin",
                        "member"
                      ]
                    },
                    "version": {
                      "type": "integer",
                      "minimum": 0,
                      "exclusiveMinimum": true
                    }
                  },
                  "required": [
                    "displayName",
                    "id",
                    "joinedAt",
                    "role",
                    "version"
                  ],
                  "additionalProperties": false
                }
              },
              "space": {
                "type": "object",
                "properties": {
                  "defaultLocale": {
                    "type": "string",
                    "enum": [
                      "ca",
                      "es",
                      "fr",
                      "en",
                      "it",
                      "pt-PT",
                      "nl",
                      "de"
                    ]
                  },
                  "id": {
                    "type": "string",
                    "pattern": "^[0-9A-HJKMNP-TV-Z]{26}$"
                  },
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 120
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "owner",
                      "admin",
                      "member"
                    ]
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "personal",
                      "couple",
                      "group"
                    ]
                  },
                  "version": {
                    "type": "integer",
                    "minimum": 0,
                    "exclusiveMinimum": true
                  }
                },
                "required": [
                  "defaultLocale",
                  "id",
                  "name",
                  "role",
                  "type",
                  "version"
                ],
                "additionalProperties": false
              }
            },
            "required": [
              "members",
              "space"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "data"
        ],
        "additionalProperties": false
      },
      "CreateSpaceRequest": {
        "type": "object",
        "properties": {
          "defaultLocale": {
            "type": "string",
            "enum": [
              "ca",
              "es",
              "fr",
              "en",
              "it",
              "pt-PT",
              "nl",
              "de"
            ]
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "type": {
            "type": "string",
            "enum": [
              "couple",
              "group"
            ]
          }
        },
        "required": [
          "defaultLocale",
          "name",
          "type"
        ],
        "additionalProperties": false
      },
...
    },
...
  },
  "paths": {
...
    "/api/v1/me/bootstrap": {
      "get": {
        "operationId": "getBootstrap",
        "tags": [
          "Session"
        ],
        "summary": "Create or load the authenticated user's private bootstrap state",
        "security": [
          {
            "FirebaseBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "The authenticated user, active memberships, and runtime feature state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BootstrapResponse"
                }
              }
            }
          },
          "401": {
            "description": "A valid Firebase ID token is required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/me": {
      "patch": {
        "operationId": "updateProfile",
        "tags": [
          "Session"
        ],
        "summary": "Update the authenticated profile or active Space",
        "security": [
          {
            "FirebaseBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated bootstrap state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BootstrapResponse"
                }
              }
            }
          },
          "400": {
            "description": "The profile update is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "A valid Firebase ID token is required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "The requested active Space is not available to this user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/spaces": {
      "post": {
        "operationId": "createSpace",
        "tags": [
          "Spaces"
        ],
        "summary": "Create a couple or group Space",
        "security": [
          {
            "FirebaseBearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{43}$"
            },
            "required": true,
            "in": "header",
            "name": "Idempotency-Key"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSpaceRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The Space and its owner membership were created.",
            "headers": {
              "Location": {
                "description": "The new Space resource.",
                "schema": {
                  "type": "string"
                }
              },
              "Idempotency-Replayed": {
                "description": "True when a prior successful command was replayed.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpaceDetailResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "User unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Idempotency conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/invitations/{token}/accept": {
      "post": {
        "operationId": "acceptSpaceInvitation",
        "tags": [
          "Invitations"
        ],
        "summary": "Accept a valid invitation idempotently",
        "security": [
          {
            "FirebaseBearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{43}$"
            },
            "required": true,
            "in": "path",
            "name": "token"
          }
        ],
        "responses": {
          "200": {
            "description": "Updated session with the accepted Space active.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BootstrapResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Invitation unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  }
}